Added two quarters to database using script!
The version two script was able to add the Spring Quarter 14-15 and the Winter Quarter 14-15 to the courses.db. (Notably, they were done on two different runs, with STOP_QUARTER being the spring quarter the first time and the winter quarter the next. This means that in the future, this script can be reused when more terms are added to Drexel One.) Note, however, that subjects did not match their database values and still have to be changed.